pyproject.toml
[project]
name = "cdnow-customer-value"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "gregorywmorris", email = "gwm1980@gmail.com"},
]
dependencies = [
    "ipykernel>=6.29.4",
    "matplotlib>=3.9.0",
    "plotly>=5.22.0",
    "pandas>=2.2.2",
    "statsmodels>=0.14.2",
    "xgboost>=2.0.3",
    "sqlalchemy>=2.0.30",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "plydata>=0.4.3",
    "plotnine>=0.13.6",
    "mizani>=0.11.4",
    "scikit-learn>=1.5.0",
    "sklearn-pandas>=2.2.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "NONE"}


[tool.pdm]
distribution = false
